Make dep on `parameterized-trigger` optional
