+++ src/main/java/org/jenkinsci/plugins/parallel_test_executor/RunListenerImpl.java
@@ -1,18 +1,18 @@
 package org.jenkinsci.plugins.parallel_test_executor;
 
-import hudson.Extension;
 import hudson.model.AbstractBuild;
 import hudson.model.TaskListener;
 import hudson.model.listeners.RunListener;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
+import org.jenkinsci.plugins.variant.OptionalExtension;
 
 /**
  * Looks for {@link TestCollector} in the build and collects the test reports.
  *
  * @author Kohsuke Kawaguchi
  */
-@Extension
+@OptionalExtension(requirePlugins = "parameterized-trigger")
 public class RunListenerImpl extends RunListener<AbstractBuild<?,?>> {
     @Override
     public void onCompleted(AbstractBuild<?,?> build, @NonNull TaskListener listener) {
