+++ src/test/java/org/jenkinsci/plugins/parallel_test_executor/ParallelTestExecutorUnitTest.java
@@ -117,7 +117,7 @@ public class ParallelTestExecutorUnitTest {
         testResult.tally();
         when(action.getResult()).thenReturn(testResult);
 
-        List<InclusionExclusionPattern> splits = ParallelTestExecutor.findTestSplits(parallelism, testMode, build, listener, false, null, null);
+        List<InclusionExclusionPattern> splits = Splitter.findTestSplits(parallelism, testMode, build, listener, false, null, null);
         assertEquals(expectedSplitSize, splits.size());
         for (InclusionExclusionPattern split : splits) {
             assertFalse(split.isIncludes());
@@ -132,7 +132,7 @@ public class ParallelTestExecutorUnitTest {
         when(action.getResult()).thenReturn(testResult);
 
         CountDrivenParallelism parallelism = new CountDrivenParallelism(5);
-        List<InclusionExclusionPattern> splits = ParallelTestExecutor.findTestSplits(parallelism, null, build, listener, false, null, null);
+        List<InclusionExclusionPattern> splits = Splitter.findTestSplits(parallelism, null, build, listener, false, null, null);
         assertEquals(2, splits.size());
         for (InclusionExclusionPattern split : splits) {
             assertFalse(split.isIncludes());
@@ -145,7 +145,7 @@ public class ParallelTestExecutorUnitTest {
         testResult.tally();
         when(action.getResult()).thenReturn(testResult);
         CountDrivenParallelism parallelism = new CountDrivenParallelism(3);
-        List<InclusionExclusionPattern> splits = ParallelTestExecutor.findTestSplits(parallelism, new JavaTestCaseName(), build, listener, false, null, null);
+        List<InclusionExclusionPattern> splits = Splitter.findTestSplits(parallelism, new JavaTestCaseName(), build, listener, false, null, null);
         assertEquals(3, splits.size());
         var allSplits = splits.stream().flatMap(s -> s.getList().stream()).collect(Collectors.toSet());
         assertThat(allSplits, hasSize(20));
@@ -169,7 +169,7 @@ public class ParallelTestExecutorUnitTest {
         testResult.tally();
         when(action.getResult()).thenReturn(testResult);
 
-        List<InclusionExclusionPattern> splits = ParallelTestExecutor.findTestSplits(parallelism, testMode, build, listener, true, null, null);
+        List<InclusionExclusionPattern> splits = Splitter.findTestSplits(parallelism, testMode, build, listener, true, null, null);
         assertEquals(expectedSplitSize, splits.size());
         List<String> exclusions = new ArrayList<>(splits.get(0).getList());
         List<String> inclusions = new ArrayList<>();
@@ -189,7 +189,7 @@ public class ParallelTestExecutorUnitTest {
     @Test
     public void findTestInJavaProjectDirectory() throws InterruptedException {
         CountDrivenParallelism parallelism = new CountDrivenParallelism(5);
-        List<InclusionExclusionPattern> splits = ParallelTestExecutor.findTestSplits(parallelism, null, build, listener, true, null, new FilePath(scanner.getBasedir()));
+        List<InclusionExclusionPattern> splits = Splitter.findTestSplits(parallelism, null, build, listener, true, null, new FilePath(scanner.getBasedir()));
         assertEquals(5, splits.size());
     }
 
@@ -207,7 +207,7 @@ public class ParallelTestExecutorUnitTest {
         expectedTests.add("FourthTest");
         expectedTests.add("FifthTest");
         assertEquals("Result does not contains expected tests.", expectedTests, data.keySet());
-        List<InclusionExclusionPattern> splits = ParallelTestExecutor.findTestSplits(parallelism, null, build, listener, true, null, new FilePath(scanner.getBasedir()));
+        List<InclusionExclusionPattern> splits = Splitter.findTestSplits(parallelism, null, build, listener, true, null, new FilePath(scanner.getBasedir()));
         assertEquals(5, splits.size());
     }
 
@@ -226,6 +226,6 @@ public class ParallelTestExecutorUnitTest {
         testResult.tally();
         when(action.getResult()).thenReturn(testResult);
 
-        assertNotNull(ParallelTestExecutor.getTestResult(project, previousBuild, listener));
+        assertNotNull(Splitter.getTestResult(project, previousBuild, listener));
     }
 }
